Implement transitive build failure caching
This is a bit of a hack and should really be implemented in nix instead.
Its incredibly useful for bisecting though, so this works for now.

Unfortunately now the `nix build` output won't be printed to stdout
since `nix build` will suppress that output when it detects
non-interactive use. I'll fix that in a follow-up.