Cache full failure logs for skip caching
`nix log` is not reliable. But we can trust the logs we produced
ourselves. Caching those already helps a great bit. If we can trust the
log, we can be sure that the failure_line does not occur in it and do
not need to rebuild to check for skips.